LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-principal-box-expected.txt


FAIL anchor-name should only apply to elements which generate a principal box assert_equals: #anchored is positioned against #inner expected 100 but got 0
